+++ command/console.go
@@ -1,16 +1,9 @@
 package command
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
-	"io"
-	"net/http"
-	"net/url"
-	"strings"
 
-	"github.com/aws/aws-sdk-go/aws/arn"
-	"github.com/kreuzwerker/awsu/strategy"
+	"github.com/kreuzwerker/awsu/console"
 	"github.com/skratchdot/open-golang/open"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
@@ -25,88 +18,23 @@ var consoleCmd = &cobra.Command{
 	},
 	RunE: func(cmd *cobra.Command, args []string) error {
 
-		creds, err := strategy.Apply(&conf)
+		cons, err := console.New(&conf)
 
 		if err != nil {
 			return err
 		}
 
-		profile := conf.Profiles[conf.Profile]
+		link, err := cons.Link()
 
-		// TODO: move the whole logic to a "console" package
-		if profile.ExternalID == "" {
-
-			a, err := arn.Parse(profile.RoleARN)
-
-			if err != nil {
-				return err
-			}
-
-			url := fmt.Sprintf("https://signin.aws.amazon.com/switchrole?account=%s&roleName=%s&displayName=%s",
-				a.AccountID,
-				strings.TrimPrefix(a.Resource, "role/"),
-				profile.Name)
-
-			if conf.Console.Open {
-				return open.Run(url)
-			}
-
-			fmt.Println(url)
-
-		} else {
-
-			fep := map[string]string{
-				"sessionId":    creds.AccessKeyID,
-				"sessionKey":   creds.SessionToken,
-				"sessionToken": creds.SessionToken,
-			}
-
-			enc, err := json.Marshal(fep)
-
-			if err != nil {
-				return fmt.Errorf("error while marshaling federation session: %s", err)
-			}
-
-			url := fmt.Sprintf("https://signin.aws.amazon.com/federation?Action=getSigninToken&Session=%s", string(url.QueryEscape(string(enc))))
-
-			var buf = bytes.NewBuffer(nil)
-
-			res, err := http.Get(url)
-
-			if err != nil {
-				return fmt.Errorf("error while requesting federation: %s", err)
-			}
-
-			defer res.Body.Close()
-
-			if _, err := io.Copy(buf, res.Body); err != nil {
-				return fmt.Errorf("error while receiving federation response body: %s", err)
-			}
-
-			var body map[string]string
-
-			if err := json.Unmarshal(buf.Bytes(), &body); err != nil {
-				return fmt.Errorf("error while unmarshaling sign-in token: %s", err)
-			}
-
-			var (
-				destination = "https://console.aws.amazon.com/"
-				issuer      = ""
-				token       = body["SigninToken"]
-			)
-
-			url = fmt.Sprintf("https://signin.aws.amazon.com/federation?Action=login&Issuer=%s&Destination=%s&SigninToken=%s\n",
-				issuer,
-				destination,
-				token)
-
-			if conf.Console.Open {
-				return open.Run(url)
-			}
-
-			fmt.Println(url)
-
+		if err != nil {
+			return err
 		}
+
+		if conf.Console.Open {
+			return open.Run(link)
+		}
+
+		fmt.Println(link)
 
 		return nil